Update README.md
Added some detail on how to narrow the scope, to get a more feasible project solution. - Focus on technique details, specific for your instrument. - This also allows for built-in short general examples in the solution.

Added further comment (than before) on legislation varying on immaterial properties rights, and this in several ways needing to be considered.